feat: hold by partition facet
