+++ packages/ats/contracts/scripts/domain/loanPortfolio/createConfiguration.ts
@@ -82,7 +82,7 @@ const LOANS_PORTFOLIO_FACETS = [
   // Hold
   "HoldReadFacet",
   "HoldManagementFacet",
-  "HoldTokenHolderFacet",
+  "HoldByPartitionFacet",
 
   // Clearing & Settlement
   "ClearingTransferFacet",
